requirements.txt
asgiref==3.6.0
boto3==1.26.114
botocore==1.29.114
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==1.3.0
Django==4.2
django-allauth==0.54.0
django-filter==23.1
django-storages==1.13.2
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
stripe==5.4.0
typing_extensions==4.5.0
urllib3==1.26.15
